Added new_id parameter to card updates

+++ sql/createOrUpdateCardVariations.sql
@@ -1,34 +1,68 @@
-CREATE OR REPLACE FUNCTION createOrUpdateCardVariations() RETURNS integer AS $$
+CREATE OR REPLACE FUNCTION createOrUpdateCardVariations(
+    character varying
+) RETURNS integer AS $$
 DECLARE
+    _new_id ALIAS FOR $1;
+
     currentRow integer := 0;
     rows integer := 0;
     row RECORD;
     row2 RECORD;
 BEGIN
-    DELETE FROM cmcard_variation;
+    IF lower(_new_id) = 'null' THEN
+        DELETE FROM cmcard_variation;
 
-	SELECT count(*) INTO rows FROM cmcard c
-            LEFT JOIN cmset s ON c.cmset = s.code
-        WHERE
-            c.new_id NOT IN (SELECT cmcard_face FROM cmcard_face) AND
-            c.new_id NOT IN (SELECT cmcard_part FROM cmcard_component_part);
+        SELECT count(*) INTO rows FROM cmcard c
+                LEFT JOIN cmset s ON c.cmset = s.code
+            WHERE
+                c.new_id NOT IN (SELECT cmcard_face FROM cmcard_face) AND
+                c.new_id NOT IN (SELECT cmcard_part FROM cmcard_component_part);
 
-    RAISE NOTICE 'variations: %/%', currentRow, rows;
-    FOR row IN SELECT new_id, c.name, cmset, cmlanguage FROM cmcard c
-            LEFT JOIN cmset s ON c.cmset = s.code
-        WHERE
-            c.new_id NOT IN (SELECT cmcard_face FROM cmcard_face) AND
-            c.new_id NOT IN (SELECT cmcard_part FROM cmcard_component_part)
-        ORDER BY s.release_date, c.name
-    LOOP
+        RAISE NOTICE 'variations: %/%', currentRow, rows;
+        FOR row IN SELECT new_id, c.name, cmset, cmlanguage FROM cmcard c
+                LEFT JOIN cmset s ON c.cmset = s.code
+            WHERE
+                c.new_id NOT IN (SELECT cmcard_face FROM cmcard_face) AND
+                c.new_id NOT IN (SELECT cmcard_part FROM cmcard_component_part)
+            ORDER BY s.release_date, c.name
+        LOOP
+            FOR row2 IN SELECT new_id FROM cmcard c
+                    LEFT JOIN cmset s ON c.cmset = s.code
+                WHERE
+                    new_id <> row.new_id AND
+                    cmset = row.cmset AND
+                    c.name = row.name AND
+                    cmlanguage = row.cmlanguage
+                ORDER BY s.release_date, c.name
+            LOOP
+                INSERT INTO cmcard_variation(
+                    cmcard,
+                    cmcard_variation)
+                VALUES(
+                    row.new_id,
+                    row2.new_id);
+            END LOOP;
+
+            currentRow := currentRow + 1;
+
+            IF currentRow % 1000 = 0 THEN
+                RAISE NOTICE 'variations: %/%', currentRow, rows;
+            END IF;
+        END LOOP;
+
+    ELSE
+        DELETE FROM cmcard_variation WHERE cmcard = _new_id;
+        SELECT new_id, name, cmset, cmlanguage INTO row  FROM cmcard c
+            WHERE new_id = _new_id;
+
         FOR row2 IN SELECT new_id FROM cmcard c
                 LEFT JOIN cmset s ON c.cmset = s.code
             WHERE
                 new_id <> row.new_id AND
                 cmset = row.cmset AND
-			    c.name = row.name AND
-			    cmlanguage = row.cmlanguage
-			ORDER BY s.release_date, c.name
+                c.name = row.name AND
+                cmlanguage = row.cmlanguage
+            ORDER BY s.release_date, c.name
         LOOP
             INSERT INTO cmcard_variation(
                 cmcard,
@@ -38,12 +72,8 @@ BEGIN
                 row2.new_id);
         END LOOP;
 
-		currentRow := currentRow + 1;
-
-		IF currentRow % 1000 = 0 THEN
-            RAISE NOTICE 'variations: %/%', currentRow, rows;
-		END IF;
-    END LOOP;
+        rows := 1;
+    END IF;    
 
     RETURN rows;
 END;
